Fix race condition with vector angles
There is a very small delay in the reset function, which allows other threads to take priority. If the chassis thread takes priority during that delay, and the chassis is still in holonomic odom mode, then the vector angle will be changed to a non-zero number. The linear chassis movements require a vector angle of zero to function properly. To fix this, the pid mode is changed to LINEAR before calling the reset function. This eliminates the race condition.